[package_manager] sync with trunk

+++ scilab/modules/gui/sci_gateway/c/sci_x_dialog.c
@@ -53,7 +53,7 @@ int C2F(sci_x_dialog)(char *fname,unsigned long fname_len)
   /* Title is a default title */
   setMessageBoxTitle(messageBoxID, _("Scilab Input Value Request"));
   /* Message */
-  setMessageBoxMultiLineMessage(messageBoxID, getStringMatrixFromStack((int)labelsAdr), nbCol*nbRow);
+  setMessageBoxMultiLineMessage(messageBoxID, getStringMatrixFromStack((size_t)labelsAdr), nbCol*nbRow);
     
   if (Rhs == 2)
     {
@@ -67,7 +67,7 @@ int C2F(sci_x_dialog)(char *fname,unsigned long fname_len)
           return FALSE;
         }
 
-      setMessageBoxInitialValue(messageBoxID, getStringMatrixFromStack((int)initialValueAdr), nbCol*nbRow);
+      setMessageBoxInitialValue(messageBoxID, getStringMatrixFromStack((size_t)initialValueAdr), nbCol*nbRow);
     }
 
   /* Display it and wait for a user input */
@@ -83,9 +83,11 @@ int C2F(sci_x_dialog)(char *fname,unsigned long fname_len)
   else
     {
       userValue = getMessageBoxValue(messageBoxID);
+	  
   
       nbCol = 1;
       CreateVarFromPtr(Rhs+1, MATRIX_OF_STRING_DATATYPE, &userValueSize, &nbCol, userValue);
+      /* TO DO : delete of userValue */
       
     }
